services/recommender/requirements.txt
# Core dependencies
fastapi==0.115.12
uvicorn==0.34.3
pydantic==2.10.5

# TensorFlow and Keras 3
tensorflow==2.18.0
keras==3.8.0

# Data processing
pandas==2.2.3
numpy==1.26.4

# MongoDB
pymongo==4.13.1

# Vault
hvac==2.3.0
kubernetes==31.0.0

# Utilities
python-dotenv==1.1.0